server/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
openai==1.54.5


